pyproject.toml
[project]
name = "nonebot-plugin-acgnshow"
version = "0.2.3"
description = "Nonebot2插件，从哔哩哔哩会员购获取简易展览数据"
readme = "README.md"
requires-python = "<4.0,>=3.9"
authors = [{ name = "Asankilp", email = "asankilp@outlook.com" }]
dependencies = [
    "nonebot2[fastapi, websockets]>=2.2.0",
    "nonebot-plugin-alconna>=0.48.0",
    "nonebot-plugin-htmlrender>=0.3.2",
    "jinja2>=3.1.4",
    "aiohttp>=3.9"
]
license = { text = "MIT" }

[project.urls]
Homepage = "https://github.com/LiteyukiStudio/nonebot-plugin-acgnshow"


[tool.nonebot]
plugins = ["nonebot_plugin_acgnshow"]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" },

]


[tool.pdm]
distribution = true

[tool.pdm.version]
source = "file"
path = "nonebot_plugin_acgnshow/__init__.py"

[tool.pdm.build]
includes = []
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
